Bytecode: keep some optimization hints from lambda
Some lambda information, such as whether an allocated block is immutable,
or the kind and layout of a bigarray we are accessing, are not taken into
account when generating bytecode. This information could be useful for
Js_of_ocaml and Wasm_of_ocaml to generate better code. We insert it as an
additional section in bytecode modules and executables, similarly to what
is done for debug events.